# backend/swagger.json
{"openapi":"3.1.0","info":{"title":"OpenAPI definition","version":"v0"},"servers":[{"url":"http://localhost:8080","description":"Generated server url"}],"paths":{"/api/crops/{id}":{"get":{"tags":["crop-controller"],"operationId":"getMyCrop","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/CropResponse"}}}}}},"put":{"tags":["crop-controller"],"operationId":"updateCrop","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateCropRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/CropResponse"}}}}}},"delete":{"tags":["crop-controller"],"operationId":"deleteCrop","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK"}}}},"/api/risk/assess":{"post":{"tags":["risk-controller"],"operationId":"assessRisk","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AssessRiskRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/RiskAssessmentResponse"}}}}}}},"/api/officer/alerts/{alertId}/interventions":{"get":{"tags":["intervention-controller"],"operationId":"getForAlert","parameters":[{"name":"alertId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/InterventionResponse"}}}}}}},"post":{"tags":["intervention-controller"],"operationId":"create","parameters":[{"name":"alertId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateInterventionRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/InterventionResponse"}}}}}}},"/api/farmers/profile":{"post":{"tags":["farmer-controller"],"operationId":"createProfile","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateFarmerProfileRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/FarmerProfileResponse"}}}}}}},"/api/dev/risk/assess":{"post":{"tags":["development-risk-controller"],"operationId":"assessRisk_1","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AssessRiskRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/RiskAssessmentResponse"}}}}}}},"/api/crops":{"get":{"tags":["crop-controller"],"operationId":"getMyCrops","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/CropResponse"}}}}}}},"post":{"tags":["crop-controller"],"operationId":"createCrop","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateCropRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/CropResponse"}}}}}}},"/api/auth/register":{"post":{"tags":["auth-controller"],"operationId":"register","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RegistrationRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/AuthenticationResponse"}}}}}}},"/api/auth/login":{"post":{"tags":["auth-controller"],"operationId":"login","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/AuthenticationResponse"}}}}}}},"/api/advisories/generate":{"post":{"tags":["advisory-controller"],"operationId":"generateAdvisory","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/GenerateAdvisoryRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/AdvisoryResponse"}}}}}}},"/api/officer/interventions/{id}":{"get":{"tags":["intervention-controller"],"operationId":"getById","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/InterventionResponse"}}}}}},"patch":{"tags":["intervention-controller"],"operationId":"update","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateInterventionRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/InterventionResponse"}}}}}}},"/api/officer/alerts/{id}/resolve":{"patch":{"tags":["distress-alert-controller"],"operationId":"resolve","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ResolveAlertRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/DistressAlertResponse"}}}}}}},"/api/officer/alerts/{id}/assign":{"patch":{"tags":["officer-alert-controller"],"operationId":"assign","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AssignAlertRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/DistressAlertResponse"}}}}}}},"/api/officer/alerts/{id}/acknowledge":{"patch":{"tags":["distress-alert-controller"],"operationId":"acknowledge","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AcknowledgeAlertRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/DistressAlertResponse"}}}}}}},"/api/notifications/{id}/read":{"patch":{"tags":["notification-controller"],"operationId":"markAsRead","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/NotificationResponse"}}}}}}},"/api/weather/forecast":{"get":{"tags":["weather-controller"],"operationId":"getForecast","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/WeatherForecastResponse"}}}}}}},"/api/weather/current":{"get":{"tags":["weather-controller"],"operationId":"getCurrentWeather","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/CurrentWeatherResponse"}}}}}}},"/api/officer/dashboard":{"get":{"tags":["officer-dashboard-controller"],"operationId":"getDashboard","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/OfficerDashboardResponse"}}}}}}},"/api/officer/alerts":{"get":{"tags":["distress-alert-controller"],"operationId":"getOfficerAlerts","parameters":[{"name":"status","in":"query","required":false,"schema":{"type":"string","enum":["OPEN","ACKNOWLEDGED","RESOLVED"]}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/DistressAlertResponse"}}}}}}}},"/api/officer/alerts/{id}":{"get":{"tags":["distress-alert-controller"],"operationId":"getOfficerAlert","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/DistressAlertResponse"}}}}}}},"/api/notifications":{"get":{"tags":["notification-controller"],"operationId":"getAll","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/NotificationResponse"}}}}}}}},"/api/notifications/{id}":{"get":{"tags":["notification-controller"],"operationId":"getById_1","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/NotificationResponse"}}}}}}},"/api/notifications/unread":{"get":{"tags":["notification-controller"],"operationId":"getUnread","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/NotificationResponse"}}}}}}}},"/api/market/prices":{"get":{"tags":["market-controller"],"operationId":"getPrices","parameters":[{"name":"cropName","in":"query","required":false,"schema":{"type":"string"}},{"name":"district","in":"query","required":false,"schema":{"type":"string"}},{"name":"state","in":"query","required":false,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/MarketPriceResponse"}}}}}}}},"/api/market/prices/history":{"get":{"tags":["market-controller"],"operationId":"getPriceHistory","parameters":[{"name":"cropName","in":"query","required":false,"schema":{"type":"string"}},{"name":"state","in":"query","required":false,"schema":{"type":"string"}},{"name":"startDate","in":"query","required":false,"schema":{"type":"string","format":"date"}},{"name":"endDate","in":"query","required":false,"schema":{"type":"string","format":"date"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/MarketPriceResponse"}}}}}}}},"/api/market/prices/compare":{"get":{"tags":["market-controller"],"operationId":"getPriceComparison","parameters":[{"name":"cropName","in":"query","required":true,"schema":{"type":"string"}},{"name":"state","in":"query","required":false,"schema":{"type":"string"}},{"name":"district","in":"query","required":false,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/MarketPriceResponse"}}}}}}}},"/api/farmers/me":{"get":{"tags":["farmer-controller"],"operationId":"getMyProfile","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/FarmerProfileResponse"}}}}}}},"/api/farmers/me/alerts":{"get":{"tags":["distress-alert-controller"],"operationId":"getFarmerAlerts","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/DistressAlertResponse"}}}}}}}},"/api/farmers/me/alerts/{id}":{"get":{"tags":["distress-alert-controller"],"operationId":"getFarmerAlert","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/DistressAlertResponse"}}}}}}},"/api/dashboard":{"get":{"tags":["dashboard-controller"],"operationId":"getDashboard_1","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/DashboardResponse"}}}}}}},"/api/advisories":{"get":{"tags":["advisory-controller"],"operationId":"getMyAdvisories","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/AdvisoryResponse"}}}}}}}},"/api/advisories/{id}":{"get":{"tags":["advisory-controller"],"operationId":"getMyAdvisory","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/AdvisoryResponse"}}}}}}}},"components":{"schemas":{"CreateCropRequest":{"type":"object","properties":{"cropName":{"type":"string","maxLength":100,"minLength":0},"cropStage":{"type":"string","maxLength":100,"minLength":0},"sowingDate":{"type":"string","format":"date"},"expectedHarvestDate":{"type":"string","format":"date"}},"required":["cropName","expectedHarvestDate","sowingDate"]},"CropResponse":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"farmerId":{"type":"integer","format":"int64"},"cropName":{"type":"string"},"cropStage":{"type":"string"},"sowingDate":{"type":"string","format":"date"},"expectedHarvestDate":{"type":"string","format":"date"},"createdAt":{"type":"string","format":"date-time"}}},"AssessRiskRequest":{"type":"object","properties":{"cropId":{"type":"integer","format":"int64"}},"required":["cropId"]},"RiskAssessmentResponse":{"type":"object","properties":{"cropId":{"type":"integer","format":"int64"},"cropName":{"type":"string"},"cropStage":{"type":"string"},"riskScore":{"type":"integer","format":"int32"},"riskLevel":{"type":"string"},"factors":{"type":"array","items":{"$ref":"#/components/schemas/RiskFactor"}},"recommendedAction":{"type":"string"},"assessedAt":{"type":"string","format":"date-time"}}},"RiskFactor":{"type":"object","properties":{"type":{"type":"string"},"severity":{"type":"string"},"contribution":{"type":"integer","format":"int32"},"reason":{"type":"string"}}},"CreateInterventionRequest":{"type":"object","properties":{"type":{"type":"string","enum":["FIELD_VISIT","PHONE_CALL","INPUT_SUPPORT","REFERRAL","OTHER"]},"description":{"type":"string","maxLength":4000,"minLength":0}},"required":["description","type"]},"InterventionResponse":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"distressAlertId":{"type":"integer","format":"int64"},"farmerId":{"type":"integer","format":"int64"},"cropId":{"type":"integer","format":"int64"},"cropName":{"type":"string"},"officerId":{"type":"integer","format":"int64"},"type":{"type":"string","enum":["FIELD_VISIT","PHONE_CALL","INPUT_SUPPORT","REFERRAL","OTHER"]},"description":{"type":"string"},"status":{"type":"string","enum":["PLANNED","IN_PROGRESS","COMPLETED","CANCELLED"]},"createdAt":{"type":"string","format":"date-time"},"updatedAt":{"type":"string","format":"date-time"},"completedAt":{"type":"string","format":"date-time"}}},"CreateFarmerProfileRequest":{"type":"object","properties":{"district":{"type":"string","maxLength":100,"minLength":0},"state":{"type":"string","maxLength":100,"minLength":0},"latitude":{"type":"number","format":"double","maximum":90.0,"minimum":-90.0},"longitude":{"type":"number","format":"double","maximum":180.0,"minimum":-180.0},"landArea":{"type":"number","format":"double","minimum":0.0}},"required":["district","state"]},"FarmerProfileResponse":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"name":{"type":"string"},"email":{"type":"string"},"phone":{"type":"string"},"role":{"type":"string","enum":["FARMER","OFFICER","ADMIN"]},"district":{"type":"string"},"state":{"type":"string"},"latitude":{"type":"number","format":"double"},"longitude":{"type":"number","format":"double"},"landArea":{"type":"number","format":"double"}}},"RegistrationRequest":{"type":"object","properties":{"name":{"type":"string","maxLength":100,"minLength":0},"email":{"type":"string","format":"email","maxLength":255,"minLength":0},"password":{"type":"string","maxLength":100,"minLength":8},"phone":{"type":"string","maxLength":20,"minLength":0}},"required":["email","name","password"]},"AuthenticationResponse":{"type":"object","properties":{"token":{"type":"string"},"userId":{"type":"integer","format":"int64"},"name":{"type":"string"},"email":{"type":"string"},"role":{"type":"string","enum":["FARMER","OFFICER","ADMIN"]}}},"LoginRequest":{"type":"object","properties":{"email":{"type":"string","format":"email","minLength":1},"password":{"type":"string","minLength":1}},"required":["email","password"]},"GenerateAdvisoryRequest":{"type":"object","properties":{"cropId":{"type":"integer","format":"int64"}},"required":["cropId"]},"AdvisoryRecommendation":{"type":"object","properties":{"category":{"type":"string"},"severity":{"type":"string"},"title":{"type":"string"},"recommendation":{"type":"string"},"reason":{"type":"string"}}},"AdvisoryResponse":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"cropId":{"type":"integer","format":"int64"},"cropName":{"type":"string"},"cropStage":{"type":"string"},"generatedAt":{"type":"string","format":"date-time"},"recommendations":{"type":"array","items":{"$ref":"#/components/schemas/AdvisoryRecommendation"}}}},"UpdateInterventionRequest":{"type":"object","properties":{"status":{"type":"string","enum":["PLANNED","IN_PROGRESS","COMPLETED","CANCELLED"]},"description":{"type":"string","maxLength":4000,"minLength":0}},"required":["status"]},"ResolveAlertRequest":{"type":"object","properties":{"officerNote":{"type":"string","maxLength":2000,"minLength":0}},"required":["officerNote"]},"DistressAlertResponse":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"farmerId":{"type":"integer","format":"int64"},"cropId":{"type":"integer","format":"int64"},"cropName":{"type":"string"},"riskScore":{"type":"integer","format":"int32"},"riskLevel":{"type":"string"},"dominantFactor":{"type":"string"},"factors":{"type":"array","items":{"$ref":"#/components/schemas/RiskFactor"}},"recommendedAction":{"type":"string"},"status":{"type":"string","enum":["OPEN","ACKNOWLEDGED","RESOLVED"]},"assignedOfficerId":{"type":"integer","format":"int64"},"officerNote":{"type":"string"},"createdAt":{"type":"string","format":"date-time"},"acknowledgedAt":{"type":"string","format":"date-time"},"resolvedAt":{"type":"string","format":"date-time"}}},"AssignAlertRequest":{"type":"object","properties":{"officerId":{"type":"integer","format":"int64"}},"required":["officerId"]},"AcknowledgeAlertRequest":{"type":"object","properties":{"note":{"type":"string","maxLength":2000,"minLength":0}},"required":["note"]},"NotificationResponse":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"distressAlertId":{"type":"integer","format":"int64"},"interventionId":{"type":"integer","format":"int64"},"farmerId":{"type":"integer","format":"int64"},"cropId":{"type":"integer","format":"int64"},"cropName":{"type":"string"},"type":{"type":"string","enum":["DISTRESS_ALERT_CREATED","DISTRESS_ALERT_ACKNOWLEDGED","INTERVENTION_CREATED","INTERVENTION_COMPLETED","INTERVENTION_CANCELLED"]},"title":{"type":"string"},"message":{"type":"string"},"status":{"type":"string","enum":["UNREAD","READ"]},"createdAt":{"type":"string","format":"date-time"},"readAt":{"type":"string","format":"date-time"}}},"CurrentWeatherResponse":{"type":"object","properties":{"timestamp":{"type":"string"},"timezone":{"type":"string"},"temperature":{"type":"number","format":"double"},"relativeHumidity":{"type":"number","format":"double"},"precipitation":{"type":"number","format":"double"},"windSpeed":{"type":"number","format":"double"},"weatherCode":{"type":"integer","format":"int32"}}},"DailyForecast":{"type":"object","properties":{"timestamps":{"type":"array","items":{"type":"string"}},"weatherCode":{"type":"array","items":{"type":"integer","format":"int32"}},"temperatureMax":{"type":"array","items":{"type":"number","format":"double"}},"temperatureMin":{"type":"array","items":{"type":"number","format":"double"}},"precipitationSum":{"type":"array","items":{"type":"number","format":"double"}},"precipitationProbabilityMax":{"type":"array","items":{"type":"number","format":"double"}},"windSpeedMax":{"type":"array","items":{"type":"number","format":"double"}},"evapotranspiration":{"type":"array","items":{"type":"number","format":"double"}}}},"HourlyForecast":{"type":"object","properties":{"timestamps":{"type":"array","items":{"type":"string"}},"temperature":{"type":"array","items":{"type":"number","format":"double"}},"relativeHumidity":{"type":"array","items":{"type":"number","format":"double"}},"precipitationProbability":{"type":"array","items":{"type":"number","format":"double"}},"precipitation":{"type":"array","items":{"type":"number","format":"double"}},"windSpeed":{"type":"array","items":{"type":"number","format":"double"}},"weatherCode":{"type":"array","items":{"type":"integer","format":"int32"}}}},"WeatherForecastResponse":{"type":"object","properties":{"timezone":{"type":"string"},"current":{"$ref":"#/components/schemas/CurrentWeatherResponse"},"hourly":{"$ref":"#/components/schemas/HourlyForecast"},"daily":{"$ref":"#/components/schemas/DailyForecast"}}},"AlertSummary":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"farmerId":{"type":"integer","format":"int64"},"cropId":{"type":"integer","format":"int64"},"cropName":{"type":"string"},"riskScore":{"type":"integer","format":"int32"},"riskLevel":{"type":"string"},"dominantFactor":{"type":"string"},"recommendedAction":{"type":"string"},"status":{"type":"string"},"assignedOfficerId":{"type":"integer","format":"int64"},"createdAt":{"type":"string","format":"date-time"}}},"InterventionSummary":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"distressAlertId":{"type":"integer","format":"int64"},"farmerId":{"type":"integer","format":"int64"},"cropId":{"type":"integer","format":"int64"},"cropName":{"type":"string"},"officerId":{"type":"integer","format":"int64"},"type":{"type":"string"},"description":{"type":"string"},"status":{"type":"string"},"createdAt":{"type":"string","format":"date-time"},"completedAt":{"type":"string","format":"date-time"}}},"OfficerDashboardResponse":{"type":"object","properties":{"summary":{"$ref":"#/components/schemas/Summary"},"recentOpenAlerts":{"type":"array","items":{"$ref":"#/components/schemas/AlertSummary"}},"recentAcknowledgedAlerts":{"type":"array","items":{"$ref":"#/components/schemas/AlertSummary"}},"activeInterventions":{"type":"array","items":{"$ref":"#/components/schemas/InterventionSummary"}},"recentlyCompletedInterventions":{"type":"array","items":{"$ref":"#/components/schemas/InterventionSummary"}}}},"Summary":{"type":"object","properties":{"openAlerts":{"type":"integer","format":"int64"},"acknowledgedAlerts":{"type":"integer","format":"int64"},"resolvedAlerts":{"type":"integer","format":"int64"},"activeInterventions":{"type":"integer","format":"int64"},"completedInterventions":{"type":"integer","format":"int64"}}},"MarketPriceResponse":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"state":{"type":"string"},"district":{"type":"string"},"market":{"type":"string"},"commodity":{"type":"string"},"variety":{"type":"string"},"grade":{"type":"string"},"arrivalDate":{"type":"string","format":"date"},"minPrice":{"type":"number","format":"double"},"maxPrice":{"type":"number","format":"double"},"modalPrice":{"type":"number","format":"double"},"unit":{"type":"string"},"currency":{"type":"string"}}},"AdvisorySummary":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"cropId":{"type":"integer","format":"int64"},"cropName":{"type":"string"},"cropStage":{"type":"string"},"generatedAt":{"type":"string","format":"date-time"},"recommendationCount":{"type":"integer","format":"int32"}}},"CropSummary":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"cropName":{"type":"string"},"cropStage":{"type":"string"},"sowingDate":{"type":"string","format":"date"},"expectedHarvestDate":{"type":"string","format":"date"},"createdAt":{"type":"string","format":"date-time"}}},"DashboardResponse":{"type":"object","properties":{"farmer":{"$ref":"#/components/schemas/FarmerSummary"},"statistics":{"$ref":"#/components/schemas/SummaryStatistics"},"recentCrops":{"type":"array","items":{"$ref":"#/components/schemas/CropSummary"}},"recentAlerts":{"type":"array","items":{"$ref":"#/components/schemas/AlertSummary"}},"recentAdvisories":{"type":"array","items":{"$ref":"#/components/schemas/AdvisorySummary"}},"recentNotifications":{"type":"array","items":{"$ref":"#/components/schemas/NotificationSummary"}},"marketSummaries":{"type":"array","items":{"$ref":"#/components/schemas/MarketSummary"}}}},"FarmerSummary":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"name":{"type":"string"},"district":{"type":"string"},"state":{"type":"string"}}},"MarketSummary":{"type":"object","properties":{"cropName":{"type":"string"},"bestMarket":{"type":"string"},"bestModalPrice":{"type":"number"},"unit":{"type":"string"},"currency":{"type":"string"},"marketsTracked":{"type":"integer","format":"int32"},"latestObservation":{"type":"string","format":"date"}}},"NotificationSummary":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"type":{"type":"string"},"title":{"type":"string"},"message":{"type":"string"},"status":{"type":"string"},"createdAt":{"type":"string","format":"date-time"},"readAt":{"type":"string","format":"date-time"},"distressAlertId":{"type":"integer","format":"int64"},"interventionId":{"type":"integer","format":"int64"}}},"SummaryStatistics":{"type":"object","properties":{"totalCrops":{"type":"integer","format":"int64"},"openDistressAlerts":{"type":"integer","format":"int64"},"acknowledgedDistressAlerts":{"type":"integer","format":"int64"},"resolvedDistressAlerts":{"type":"integer","format":"int64"},"activeInterventions":{"type":"integer","format":"int64"},"unreadNotifications":{"type":"integer","format":"int64"},"totalAdvisories":{"type":"integer","format":"int64"}}}}}}